MAINT: special: wrap `exp`, `log` funcs from xsf
[skip cirrus]
