+++ scipy/special/_special_ufuncs.cpp
@@ -13,6 +13,7 @@
 #include "xsf/digamma.h"
 #include "xsf/ellip.h"
 #include "xsf/erf.h"
+#include "xsf/exp.h"
 #include "xsf/expint.h"
 #include "xsf/fresnel.h"
 #include "xsf/gamma.h"
@@ -21,6 +22,7 @@
 #include "xsf/kelvin.h"
 #include "xsf/lambertw.h"
 #include "xsf/legendre.h"
+#include "xsf/log.h"
 #include "xsf/log_exp.h"
 #include "xsf/mathieu.h"
 #include "xsf/par_cyl.h"
@@ -45,6 +47,7 @@
 
 extern const char *_cospi_doc;
 extern const char *_sinpi_doc;
+extern const char *_log1pmx_doc;
 extern const char *airy_doc;
 extern const char *airye_doc;
 extern const char *bei_doc;
@@ -71,6 +74,9 @@ extern const char *erfc_doc;
 extern const char *erfcx_doc;
 extern const char *erfi_doc;
 extern const char *exp1_doc;
+extern const char *expm1_doc;
+extern const char *exp2_doc;
+extern const char *exp10_doc;
 extern const char *expi_doc;
 extern const char *expit_doc;
 extern const char *exprel_doc;
@@ -119,6 +125,7 @@ extern const char *k1e_doc;
 extern const char *kv_doc;
 extern const char *kve_doc;
 extern const char *lambertw_doc;
+extern const char *log1p_doc;
 extern const char *logit_doc;
 extern const char *loggamma_doc;
 extern const char *log_expit_doc;
@@ -174,6 +181,8 @@ extern const char *tandg_doc;
 extern const char *voigt_profile_doc;
 extern const char *wofz_doc;
 extern const char *wright_bessel_doc;
+extern const char *xlogy_doc;
+extern const char *xlog1py_doc;
 extern const char *y0_doc;
 extern const char *y1_doc;
 extern const char *yv_doc;
@@ -337,6 +346,20 @@ PyMODINIT_FUNC PyInit__special_ufuncs() {
     PyObject *exprel = xsf::numpy::ufunc(
         {static_cast<xsf::numpy::d_d>(xsf::exprel), static_cast<xsf::numpy::f_f>(xsf::exprel)}, "exprel", exprel_doc);
     PyModule_AddObjectRef(_special_ufuncs, "exprel", exprel);
+
+    PyObject *expm1 =
+        xsf::numpy::ufunc({static_cast<xsf::numpy::d_d>(xsf::expm1), static_cast<xsf::numpy::f_f>(xsf::expm1),
+                           static_cast<xsf::numpy::D_D>(xsf::expm1), static_cast<xsf::numpy::F_F>(xsf::expm1)},
+                          "expm1", expm1_doc);
+    PyModule_AddObjectRef(_special_ufuncs, "expm1", expm1);
+
+    PyObject *exp2 = xsf::numpy::ufunc(
+        {static_cast<xsf::numpy::d_d>(xsf::exp2), static_cast<xsf::numpy::f_f>(xsf::exp2)}, "exp2", exp2_doc);
+    PyModule_AddObjectRef(_special_ufuncs, "exp2", exp2);
+
+    PyObject *exp10 = xsf::numpy::ufunc(
+        {static_cast<xsf::numpy::d_d>(xsf::exp10), static_cast<xsf::numpy::f_f>(xsf::exp10)}, "exp10", exp10_doc);
+    PyModule_AddObjectRef(_special_ufuncs, "exp10", exp10);
 
     PyObject *erf = xsf::numpy::ufunc({static_cast<xsf::numpy::d_d>(xsf::erf), static_cast<xsf::numpy::f_f>(xsf::erf),
                                        static_cast<xsf::numpy::D_D>(xsf::erf), static_cast<xsf::numpy::F_F>(xsf::erf)},
@@ -612,6 +635,29 @@ PyMODINIT_FUNC PyInit__special_ufuncs() {
          static_cast<xsf::numpy::fF_F>(xsf::cyl_bessel_ke), static_cast<xsf::numpy::dD_D>(xsf::cyl_bessel_ke)},
         "kve", kve_doc);
     PyModule_AddObjectRef(_special_ufuncs, "kve", kve);
+
+    PyObject *log1p =
+        xsf::numpy::ufunc({static_cast<xsf::numpy::d_d>(xsf::log1p), static_cast<xsf::numpy::f_f>(xsf::log1p),
+                           static_cast<xsf::numpy::D_D>(xsf::log1p), static_cast<xsf::numpy::F_F>(xsf::log1p)},
+                          "log1p", log1p_doc);
+    PyModule_AddObjectRef(_special_ufuncs, "log1p", log1p);
+
+    PyObject *_log1pmx =
+        xsf::numpy::ufunc({static_cast<xsf::numpy::d_d>(xsf::log1pmx), static_cast<xsf::numpy::f_f>(xsf::log1pmx)},
+                          "_log1pmx", _log1pmx_doc);
+    PyModule_AddObjectRef(_special_ufuncs, "_log1pmx", _log1pmx);
+
+    PyObject *xlogy =
+        xsf::numpy::ufunc({static_cast<xsf::numpy::dd_d>(xsf::xlogy), static_cast<xsf::numpy::ff_f>(xsf::xlogy),
+                           static_cast<xsf::numpy::DD_D>(xsf::xlogy), static_cast<xsf::numpy::FF_F>(xsf::xlogy)},
+                          "xlogy", xlogy_doc);
+    PyModule_AddObjectRef(_special_ufuncs, "xlogy", xlogy);
+
+    PyObject *xlog1py =
+        xsf::numpy::ufunc({static_cast<xsf::numpy::dd_d>(xsf::xlog1py), static_cast<xsf::numpy::ff_f>(xsf::xlog1py),
+                           static_cast<xsf::numpy::DD_D>(xsf::xlog1py), static_cast<xsf::numpy::FF_F>(xsf::xlog1py)},
+                          "xlog1py", xlog1py_doc);
+    PyModule_AddObjectRef(_special_ufuncs, "xlog1py", xlog1py);
 
     PyObject *log_expit =
         xsf::numpy::ufunc({static_cast<xsf::numpy::d_d>(xsf::log_expit), static_cast<xsf::numpy::f_f>(xsf::log_expit),
